+++ src/execnet/rsync.py
@@ -173,7 +173,6 @@ class RSync:
             self._receivequeue.put((channel, req))
 
         channel = gateway.remote_exec(execnet.rsync_remote)
-        channel.reconfigure(py2str_as_py3str=False, py3str_as_py2str=False)
         channel.setcallback(itemcallback, endmarker=None)
         channel.send((str(destdir), options))
         self._channels[channel] = finishedcallback